+++ backend/sivujetti/src/Page/PagesController.php
@@ -111,7 +111,7 @@ final class PagesController {
                 "baseUrl" => WebPageAwareTemplate::makeUrl("/", true),
                 "assetBaseUrl" => WebPageAwareTemplate::makeUrl("/", false),
                 "pageTypes" => $theWebsite->pageTypes->getArrayCopy(),
-                "activeTheme" => (object) ["id" => $theWebsite->activeTheme],
+                "activeTheme" => (object) ["id" => $theWebsite->activeTheme->id],
                 "blockRenderers" => $apiCtx->validBlockRenderers,
                 "showGoToDashboardMode" => $config->get("app.showGoToDashboardMode", false),
                 "dashboardUrl" => $config->get("app.dashboardUrl", ""),